+++ java/integTest/src/test/java/glide/SharedClientTests.java
@@ -5,9 +5,12 @@ import static glide.TestUtilities.commonClientConfig;
 import static glide.TestUtilities.commonClusterClientConfig;
 import static glide.TestUtilities.getRandomString;
 import static glide.api.BaseClient.OK;
-import static glide.api.models.configuration.RequestRoutingConfiguration.SimpleMultiNodeRoute.ALL_PRIMARIES;
+import static glide.api.models.commands.PasswordUpdateMode.RE_AUTHENTICATE;
+import static glide.api.models.commands.PasswordUpdateMode.USE_ON_NEW_CONNECTION;
+import static glide.api.models.configuration.RequestRoutingConfiguration.SimpleMultiNodeRoute.ALL_NODES;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -15,7 +18,6 @@ import static org.junit.jupiter.api.Assertions.fail;
 import glide.api.BaseClient;
 import glide.api.GlideClient;
 import glide.api.GlideClusterClient;
-import glide.api.models.commands.PasswordUpdateMode;
 import glide.api.models.exceptions.RequestException;
 import java.util.ArrayList;
 import java.util.List;
@@ -32,7 +34,6 @@ import lombok.SneakyThrows;
 import net.bytebuddy.utility.RandomString;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -186,55 +187,83 @@ public class SharedClientTests {
     }
 
     @SneakyThrows
-    @ParameterizedTest(autoCloseArguments = false)
-    @MethodSource("getClients")
-    public void password_update(BaseClient client) {
-        if (client instanceof GlideClient) return;
+    private void setPassword(BaseClient client, String password) {
+        if (client instanceof GlideClient) {
+            assertEquals("OK", ((GlideClient) client).configSet(Map.of("requirepass", password)).get());
+        } else {
+            assertEquals(
+                    "OK", ((GlideClusterClient) client).configSet(Map.of("requirepass", password)).get());
+        }
+    }
 
+    @SneakyThrows
+    private void dropConnection(BaseClient client, boolean disconnect) {
+        var command =
+                disconnect ? new String[] {"CLIENT", "KILL", "TYPE", "NORMAL"} : new String[] {"RESET"};
+        if (client instanceof GlideClient) {
+            ((GlideClient) client).customCommand(command).get();
+        } else {
+            ((GlideClusterClient) client).customCommand(command, ALL_NODES).get();
+        }
+    }
+
+    private static Stream<Arguments> clientsAndReconnectMode() {
+        return Stream.of(
+                Arguments.of(standaloneClient, true),
+                Arguments.of(standaloneClient, false),
+                Arguments.of(clusterClient, true),
+                Arguments.of(clusterClient, false));
+    }
+
+    @SneakyThrows
+    @ParameterizedTest(autoCloseArguments = false)
+    @MethodSource("clientsAndReconnectMode")
+    public void password_update(BaseClient client, boolean reconnect) {
         var key = UUID.randomUUID().toString();
         var pwd = UUID.randomUUID().toString();
         client.set(key, "meow meow").get();
 
-        try (BaseClient testClient = client instanceof GlideClient
-                ? GlideClient.createClient(commonClientConfig().requestTimeout(2000).build()).get()
-                : GlideClusterClient.createClient(commonClusterClientConfig().requestTimeout(2000).build()).get()) {
+        try (BaseClient testClient =
+                client instanceof GlideClient
+                        ? GlideClient.createClient(commonClientConfig().build()).get()
+                        : GlideClusterClient.createClient(commonClusterClientConfig().build()).get()) {
 
             // validate that we can get the value
             assertEquals("meow meow", testClient.get(key).get());
 
             // set the password and forcefully drop connection for the second client
-            if (client instanceof GlideClient) {
-                ((GlideClient) client).configSet(Map.of("requirepass", pwd)).get();
-                ((GlideClient) client).customCommand(new String[]{"CLIENT", "KILL", "TYPE", "NORMAL"}).get();
-            } else {
-                ((GlideClusterClient) client).configSet(Map.of("requirepass", pwd)).get();
-                ((GlideClusterClient) client).customCommand(new String[]{"CLIENT", "KILL", "TYPE", "NORMAL"}, ALL_PRIMARIES).get();
-            }
+            setPassword(client, pwd);
+            dropConnection(client, reconnect);
 
             // client should reconnect, but will receive NOAUTH error
             var exception = assertThrows(ExecutionException.class, () -> testClient.get(key).get());
+            assertInstanceOf(RequestException.class, exception.getCause());
 
-            assertEquals("OK", testClient.updateConnectionPassword(pwd, PasswordUpdateMode.RE_AUTHENTICATE).get());
+            assertEquals("OK", testClient.updateConnectionPassword(pwd, RE_AUTHENTICATE).get());
 
             // after setting new password we should be able to work with the server
             assertEquals("meow meow", testClient.get(key).get());
 
             // unset the password and drop connection again
-            if (client instanceof GlideClient) {
-                ((GlideClient) client).configSet(Map.of("requirepass", "")).get();
-                ((GlideClient) client).customCommand(new String[]{"CLIENT", "KILL", "TYPE", "NORMAL"}).get();
-            } else {
-                ((GlideClusterClient) client).configSet(Map.of("requirepass", "")).get();
-                ((GlideClusterClient) client).customCommand(new String[]{"CLIENT", "KILL", "TYPE", "NORMAL"}, ALL_PRIMARIES).get();
-            }
+            setPassword(client, pwd);
+            dropConnection(client, reconnect);
 
             // client should reconnect, but will receive NOAUTH error
             exception = assertThrows(ExecutionException.class, () -> testClient.get(key).get());
+            assertInstanceOf(RequestException.class, exception.getCause());
 
-            assertEquals("OK", testClient.updateConnectionPassword(PasswordUpdateMode.USE_ON_NEW_CONNECTION).get());
+            assertEquals(
+                    "OK",
+                    testClient
+                            .updateConnectionPassword(reconnect ? USE_ON_NEW_CONNECTION : RE_AUTHENTICATE)
+                            .get());
 
             // after updating connection params we should be able to work with the server
             assertEquals("meow meow", testClient.get(key).get());
+        } finally {
+            setPassword(client, pwd);
+            client.updateConnectionPassword(USE_ON_NEW_CONNECTION).get();
+            client.updateConnectionPassword(RE_AUTHENTICATE).get();
         }
     }
 }